requirements.txt
Flask==3.0.0
openai==1.12.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
gunicorn==21.2.0